backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
mysql-connector-python==9.0.0
